Revert "update API to check status complete watch full video"
This reverts commit 8d4d4e1ff3ed876e77850ea75f6af15d65567c61.
