--- src/main/java/com/extension/tracking/exceptions/NotFoundException.java
@@ -1,14 +1,0 @@
-package com.extension.tracking.exceptions;
-
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException() {}
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-
-    public NotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
