--- src/main/java/com/extension/tracking/model/enums/WatchStatus.java
@@ -1,6 +1,0 @@
-package com.extension.tracking.model.enums;
-
-public enum WatchStatus {
-    COMPLETE,
-    INCOMPLETE
-}
