+++ src/main/java/com/extension/tracking/repositories/TrackingRepository.java
@@ -1,16 +1,8 @@
 package com.extension.tracking.repositories;
 
-import com.extension.tracking.model.entities.Tracking;
+import com.extension.tracking.entities.Tracking;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
-
 @Repository
-public interface TrackingRepository extends MongoRepository<Tracking, String> {
-    Optional<Tracking> findByVideoIdAndUserChannel(String videoId, String userChannel);
-
-    List<Tracking> findByUserChannelOrderByLastModifiedDesc(String userChannel);
-
-}
+public interface TrackingRepository extends MongoRepository<Tracking, String> {}
